Rename delays -> delay
